+++ engine/components/sql/src/main/scala/pl/touk/nussknacker/sql/db/schema/TableDefinition.scala
@@ -1,5 +1,6 @@
 package pl.touk.nussknacker.sql.db.schema
 
+import pl.touk.nussknacker.engine.api.typed.TypedObjectDefinition
 import pl.touk.nussknacker.engine.api.typed.typing.{Typed, TypedClass, TypedObjectTypingResult, TypingResult}
 
 import java.sql.ResultSetMetaData
@@ -10,6 +11,16 @@ object TableDefinition {
     TableDefinition(
       columnDefs = (1 to resultMeta.getColumnCount).map(ColumnDefinition(_, resultMeta)).toList
     )
+
+  def apply(typedObjectDefinition: TypedObjectDefinition): TableDefinition = {
+    val columnDefinitions = typedObjectDefinition.fields.zipWithIndex
+      .map { case (typing, index) =>
+        ColumnDefinition(index + 1, typing)
+      }
+    TableDefinition(
+      columnDefs = columnDefinitions.toList
+    )
+  }
 }
 
 case class TableDefinition(columnDefs: List[ColumnDefinition]) {