+++ engine/components/sql/src/test/scala/pl/touk/nussknacker/sql/service/DatabaseLookupEnricherWithCacheTest.scala
@@ -2,20 +2,20 @@ package pl.touk.nussknacker.sql.service
 
 import pl.touk.nussknacker.engine.api.typed.TypedMap
 import pl.touk.nussknacker.sql.db.query.ResultSetStrategy
-import pl.touk.nussknacker.sql.db.schema.TableDefinition
+import pl.touk.nussknacker.sql.db.schema.{MetaDataProviderFactory, TableDefinition}
 import pl.touk.nussknacker.sql.service.DatabaseQueryEnricher.CacheTTLParamName
-import pl.touk.nussknacker.sql.utils.BaseDatabaseQueryEnricherTest
+import pl.touk.nussknacker.sql.utils.BaseHsqlQueryEnricherTest
 
 import scala.concurrent.Await
 
-class DatabaseLookupEnricherWithCacheTest extends BaseDatabaseQueryEnricherTest {
+class DatabaseLookupEnricherWithCacheTest extends BaseHsqlQueryEnricherTest {
 
   import scala.collection.JavaConverters._
   import scala.concurrent.duration._
 
-  override val service = new DatabaseLookupEnricher(dbConf)
+  override val service = new DatabaseLookupEnricher(hsqlDbPoolConfig, new MetaDataProviderFactory().create(hsqlDbPoolConfig))
 
-  override val prepareDbDDLs: List[String] = List(
+  override val prepareHsqlDDLs: List[String] = List(
     "CREATE TABLE persons (id INT, name VARCHAR(40));",
     "INSERT INTO persons (id, name) VALUES (1, 'John')"
   )